+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Add checks for contact in contributors.tsv, empty columns in TSV
 - Fix shared upload check for unreferenced files
 - Fix shared upload error messaging to display correct paths
+- Moved Validator and validation_class_iter to ingest-validation-tests
+- Deprecated cleanup examples
 
 ## v0.0.39
 - Create Object-by-Analyte EPIC directory schema